Fix session cookie invalidation if downloads are opened in new tab
HTTP_ACCEPT header was used as one item of the session cookie's hash
algorithm. The problem is that our JSON-Requests set HTTP_ACCEPT to
`*/*` but when opening a link in a new browser tab HTTP_ACCEPT is
`text/html,application/xhtml+xml`. Thus hash was different and the
cookie could not be deciphered.